# Conflicts:
#	src/main/java/com/example/itspower/service/impl/ReportServiceImpl.java
#	target/classes/application.properties
#	target/classes/com/example/itspower/controller/ReportController.class
#	target/classes/com/example/itspower/controller/ViewController.class
#	target/classes/com/example/itspower/model/entity/GroupEntity.class
#	target/classes/com/example/itspower/model/entity/ReportEntity.class
#	target/classes/com/example/itspower/model/resultset/ReportDto.class
#	target/classes/com/example/itspower/model/resultset/ViewAllDto.class
#	target/classes/com/example/itspower/repository/repositoryjpa/ReportJpaRepository.class
#	target/classes/com/example/itspower/response/group/ViewDetailGroups.class
#	target/classes/com/example/itspower/response/group/ViewDetailsResponse$ViewDetailsResponseBuilder.class
#	target/classes/com/example/itspower/response/group/ViewDetailsResponse.class
#	target/classes/com/example/itspower/response/report/ReportResponse.class
#	target/classes/com/example/itspower/response/view/RestObjectResponse.class
#	target/classes/com/example/itspower/service/ReportService.class
#	target/classes/com/example/itspower/service/impl/ReportServiceImpl.class
#	target/classes/com/example/itspower/service/impl/UserServiceImpl.class
#	target/classes/com/example/itspower/service/impl/ViewDetailSerivceImpl.class

+++ src/main/java/com/example/itspower/response/group/ViewDetailsResponse.java
@@ -15,7 +15,7 @@ public class ViewDetailsResponse {
     private Integer reportDemarcation;
     private Double laborProductivity;
     private Integer partTimeNum;
-    private Integer restNum;
+    private Double restNum;
     private Integer studentNum;
     private Integer riceCus;
     private Integer riceEmp;
